requirements.txt
Django==5.0.4
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
crispy-bootstrap5==2024.2
django-crispy-forms==2.1
gunicorn==21.2.0
Pillow==10.3.0
